requirements.txt
# 核心依賴
diffusers>=0.24.0
transformers>=4.34.0
accelerate>=0.24.0
torch>=2.0.0

# 視訊與影像處理
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.23
einops>=0.6.0

# 用於命令列參數解析和進度顯示（可選）
tqdm>=4.64